Refactor transcript generation to support personalized voice and mindfulness preferences with proper formatting
Key changes:

- Added clean_escaped_newlines function to remove \n escapes while preserving \n
- Updated node_generate_transcript to accept config for voice_blueprint and mindfulness_profile
- Integrated dynamic prompt generation based on user preferences
- Improved transcript formatting for TTS compatibility
- Enhanced personalization by tailoring meditation content to user's voice characteristics and mindfulness preferences
This change ensures that generated transcripts are properly formatted for TTS systems and are personalized to each user's unique voice and mindfulness preferences.
