+++ backend/agent_b_synth/main.py
@@ -1,4 +1,3 @@
-import os
 import uvicorn
 from fastapi import FastAPI
 from a2a.server.apps import A2AFastAPIApplication
